Implemented script according to all requirements

+++ ntp_deploy.sh
@@ -1,15 +1,17 @@
 # /bin/bash
 
-DIR=$(dirname $0)
-# Not sure this is needed
-# sudo apt-get update
+DIR="$( cd "$( dirname "${BASH_SOURCE[0]}" )" && pwd )"
 
 # Install ntp server package
-sudo apt-get -y install ntp
-sudo apt-get -y install git
+if [ ! $(dpkg -s ntp | grep Status | cut -d" " -f4) == "installed" ]; then
+  sudo apt-get update
+  sudo apt-get -y install ntp
+  sudo apt-get -y install git
+fi
 
-cd eval /etc/ntp.conf
-git init && git status && git add . && git commit -m "Fixing state"
+if [ ! -d /etc/.git/ ]; then
+  cd /etc && git init && git add ntp.conf && git commit --untracked-files=no -m "Fixing state of ntp.conf"
+fi
 # replace the first default server to "ua.pool.ntp.org"
 sudo sed -i "s/0.ubuntu.pool.ntp.org/ua.pool.ntp.org/" /etc/ntp.conf
 
@@ -22,9 +24,9 @@ sudo sed -i "s/pool 3.ubuntu.pool.ntp.org iburst//" /etc/ntp.conf
 sudo /etc/init.d/ntp restart
 
 #write out current crontab
-touch mycron
+cd $DIR && touch mycron && echo "* * * * * bash $DIR/ntp_verify.sh" >> mycron && crontab mycron && rm mycron
 #echo new cron into cron file
-echo "* * * * * $DIR/ntp_verify.sh" >> mycron
+
 #install new cron file
-sudo crontab mycron
-rm mycron
+
+
